test(808): do not run golden file tests on Windows
